requirements.txt
python-pptx>=0.6.21
google-generativeai>=0.4.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0
